Move code to src
