+++ src/ClimaAtmos.jl
@@ -15,6 +15,7 @@ import .TurbulenceConvection as TC
 include("thermo_state.jl")
 
 include(joinpath("tendencies", "held_suarez.jl"))
+include(joinpath("tendencies", "vertical_diffusion_boundary_layer.jl"))
 include(joinpath("tendencies", "rayleigh_sponge.jl"))
 include(joinpath("tendencies", "viscous_sponge.jl"))
 include(joinpath("tendencies", "advection.jl"))
